Moved libFTR bindings into a package
